refactor: extract convertStories helper
